+++ CHANGELOG.md
@@ -16,6 +16,7 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 
 ### New features
 
+* Label smoothing schedule
 * Update the OpenNMT tokenizer to 1.3.0 and use its Python package instead of requiring a manual compilation (Linux only)
 * Include a catalog of models in the library package and allow model selection with the `--model_type` command line option